requirements.ldap.txt
python-ldap==3.3.1
django-auth-ldap==2.2.0
